requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
aiosqlite==0.20.0
python-dotenv==1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
